+++ src/localization/en/appErrors.ts
@@ -1,4 +1,13 @@
 export default {
   generic: "Unexpected error: {{text}}",
   fileSharingNotAvailable: "File sharing is not available",
+  dataImport: {
+    importFromMobileNotAllawed: "Data import from Arena Mobile not allowed",
+    invalidNodeInRecord:
+      'Invalid node in record "{{recordUuid}}", node "{{nodeUuid}}" with node definition "{{nodeDefName}}" (uuid "{{nodeDefUuid}}"): {{details}}',
+    noRecordsFound:
+      "No records found in the import file or wrong file format",
+    recordOwnedByAnotherUser:
+      'Cannot update record "{{recordKeyValues}}" as it is owned by another user',
+  },
 };