+++ src/localization/es/appErrors.ts
@@ -1,4 +1,14 @@
 export default {
   generic: "Error inesperado: {{text}}",
   fileSharingNotAvailable: "El uso compartido de archivos no está disponible",
+  dataImport: {
+    importFromMobileNotAllawed:
+      "Importación de datos desde Arena Mobile no permitida",
+    invalidNodeInRecord:
+      'Nodo no válido en el registro "{{recordUuid}}", nodo "{{nodeUuid}}" con definición de nodo "{{nodeDefName}}" (uuid "{{nodeDefUuid}}"): {{details}}',
+    noRecordsFound:
+      "No se encontraron registros en el archivo de importación o el formato de archivo es incorrecto",
+    recordOwnedByAnotherUser:
+      'No se puede actualizar el registro "{{recordKeyValues}}" porque pertenece a otro usuario',
+  },
 };