+++ src/localization/fi/appErrors.ts
@@ -1,4 +1,14 @@
 export default {
   generic: "Odottamaton virhe: {{text}}",
   fileSharingNotAvailable: "Tiedoston jakaminen ei ole käytettävissä",
+  dataImport: {
+    importFromMobileNotAllawed:
+      "Tietojen tuonti Arena Mobilesta ei ole sallittu",
+    invalidNodeInRecord:
+      'Virheellinen solmu tietueessa "{{recordUuid}}", solmu "{{nodeUuid}}" solmumäärittelyllä "{{nodeDefName}}" (uuid "{{nodeDefUuid}}"): {{details}}',
+    noRecordsFound:
+      "Tuontitiedostosta ei löytynyt tietueita tai tiedostomuoto on väärä",
+    recordOwnedByAnotherUser:
+      'Tietuetta "{{recordKeyValues}}" ei voi päivittää, koska se kuuluu toiselle käyttäjälle',
+  },
 };