+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.5] - TO-BE-RELEASED
+
+### Added
+
+- Allow merging records with same internal identifier (modified both client and server side);
+
+## [2.7.4] - 2026-08-23
+
+### Added
+
+- Support min count items auto-creation;
+
 ## [2.7.3] - 2026-08-14
 
 ### Fixed